chat_Bot-master/chat_Bot-master/requirements.txt
streamlit
requests
python-dotenv
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pymongo
google-generativeai
